Start refactor to use transformers.jl
