--- src/RotaryPositionalEmbedding.jl
@@ -1,3 +1,0 @@
-ϵ = 0.001
-
-RoPE(x, position) = x * exp(position * im * ϵ)
