--- test/testsymbolcreator.jl
@@ -1,8 +1,0 @@
-using SymbolicTransformer.SymbolCreator
-
-@testset "SymbolCreator.jl" begin
-    cache = Dict("test" => "value")
-    getter = SymbolCreator.TransformerLensCacheParser(cache)
-    #check that getter is a function 
-    @test @isdefined getter
-end
